WIP: Run Redfish BMC alongside IPMI BMC
Install `sushy-tools` emulator on the `bmc` machine to control
the other OpenStack instances from the undercloud.

This should allow for `redfish` based power control which should
be more reliable. On top of that, this change opens up the opportunity
to run deployments that rely on the advanced BMC features.

There are still a handful of issues with sushy-tools that we
are trying to sort out.